% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH_within_Gibbs_sampler.R
\name{bidc_pd}
\alias{bidc_pd}
\title{Bayesian Inference of Default Correlations Using Probabilities of Default}
\usage{
bidc_pd(S = 2000L, y, p_d, b = 50, verbose = 0L, init = NULL)
}
\arguments{
\item{S}{number of iterations for the chain.}

\item{y}{integer matrix of size \code{N x tau} containing a 1 in (n,t)
if the n-th client at time t defaulted, and 0 otherwise.}

\item{p_d}{matrix with probability of default of the n-th client at
time t.}

\item{b}{shape2 parameter of the beta proposal for rho. Default value is 50
tends to work well.}

\item{verbose}{integer. Function prints every \code{verbose} iterations. The
default value of 0L prints no output.}

\item{init}{an optional initialization for the chain. If \code{NULL}
(default), the M_t's are initialized from the prior and rho from a beta close
to 0.5.}
}
\value{
a matrix of size \code{S x (tau+1)}, where the first tau columns
correspond to the latent factors, and the last one contains samples for rho.
}
\description{
This functions runs an MCMC chain to perform Bayesian inference on the default
correlation parameter, by using probabilites of default estimated for each
client. See Details for more information.
}
\details{
The MCMC chain is a Metropolis-Hastings-withing-Gibbs (MHwG) sampler. The MH
updates are needed because the full conditionals of the parameters are
not analytically tractable. The prior on rho is uniform in (0,1), and for the
latent factors are standard normals.

The approach used here requires knowing the actual outcomes of the defaults
for a fixed number of clients (N) over a time period (tau), which are encoded
in the matrix \code{y}. It also requires an estimate of the probability of
default (PD) for each entry in \code{y} in a matrix we call \code{p_d}.
The inferences are better when the estimates of the PDs are better.

It is crucial to understand that client n in time t1 does not need to
correspond to the same person in the n-th row at time t2. In other words,
we assume exchangeability within the rows of y. The user only needs to
make sure that every entry in \code{p_d} is an estimate for the corresponding
entry in \code{y}.
}
